Add pulp dir restore logic
